+++ src/commons/authentication.rs
@@ -55,7 +55,8 @@ impl AuthenticationClass {
         authentication_class_name: &str,
     ) -> Result<AuthenticationClass, Error> {
         client
-            .get::<AuthenticationClass>(authentication_class_name, None) // AuthenticationClass has ClusterScope
+            .get_cluster_api::<AuthenticationClass>()
+            .get(authentication_class_name)
             .await
     }
 }